Fix failing master CI: App compile/test (react resolution) + lint
Three pre-existing failures on master, none from application logic:

1. Compile (compile-app) and App Test both failed on
   NetworkDevice/TopologyGraph.tsx "Cannot find module 'react'".
   App/Tests/Dashboard/TopologyLayout.test.ts imported the pure
   computeTopologyLayout() from the .tsx component, which also imports
   react — dragging react into the App tsc/jest compile graph, where it
   is not resolvable (react lives in the Dashboard FeatureSet, not App).
   Extract the pure, react-free layout (computeTopologyLayout + its
   helpers, TopologyPoint, clamp) into a sibling TopologyLayout.ts;
   TopologyGraph.tsx and the test now import from there. The component's
   own comment already stated this logic "is unit-tested in a plain node
   environment" — this makes that true.

2. Common Jobs (lint) failed on LLMService.ts multiline-comment-style.
   Convert the consecutive // comment block to a /* */ block (eslint
   --fix). Comment-only change; no AI/Sentinel behavior change.

Verified: App tsc compiles clean, all 113 App Dashboard tests pass
(incl. the previously-failing TopologyLayout suite), targeted lint of
the affected files is clean.
