+++ Common/Server/Utils/LLM/LLMService.ts
@@ -232,12 +232,14 @@ export default class LLMService {
       Object.assign(data, request.additionalParams);
     }
 
-    // OpenAI's newer model families (gpt-5, o1, o3, ...) reject the legacy
-    // `max_tokens` parameter and require `max_completion_tokens` instead; the
-    // two are mutually exclusive. When a provider opts into
-    // `max_completion_tokens` via additionalParams, drop the default
-    // `max_tokens` so the request is accepted. Legacy OpenAI-compatible
-    // backends (Ollama, vLLM, LocalAI, ...) keep receiving `max_tokens`.
+    /*
+     * OpenAI's newer model families (gpt-5, o1, o3, ...) reject the legacy
+     * `max_tokens` parameter and require `max_completion_tokens` instead; the
+     * two are mutually exclusive. When a provider opts into
+     * `max_completion_tokens` via additionalParams, drop the default
+     * `max_tokens` so the request is accepted. Legacy OpenAI-compatible
+     * backends (Ollama, vLLM, LocalAI, ...) keep receiving `max_tokens`.
+     */
     if (
       data["max_completion_tokens"] !== undefined &&
       data["max_tokens"] !== undefined
